+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## v0.3.34 - Ed25519 Sign With Key
+  * Commit: 
+  * Ed25519 Sign Message With Key
 ## v0.3.33 - Ed25519 Signature 
   * Commit: 2026-05-26 09:03
   * Generate Ed25519 Keys